Renamed some files and moved checkPlacement into a folder to mimic the checkMove structure.
