+++ gameCode/moveLogic/bugLogic/ant.logic.js
