+++ gameCode/moveLogic/bugLogic/beetle.logic.js
