+++ gameCode/moveLogic/bugLogic/hopper.logic.js
