+++ gameCode/moveLogic/bugLogic/queenBee.logic.js
