+++ gameCode/moveLogic/bugLogic/spider.logic.js
