+++ gameCode/moveLogic/checkMove.js
