+++ gameCode/moveLogic/continuity.util.js
