+++ gameCode/placementLogic/checkPlacement.js
